feat: make lazy column with offset reusable
